Add more tests and rename distance files

+++ load_test.go
@@ -10,7 +10,7 @@ var allTestData = []struct {
 	expected Float64Data
 }{
 	{
-		[]interface{}{1.0, "2", 3.0, 4, "4.0", 5, time.Duration(6), time.Duration(-7)},
+		[]interface{}{1.0, "2", 3.0, uint(4), "4.0", 5, time.Duration(6), time.Duration(-7)},
 		Float64Data{1.0, 2.0, 3.0, 4.0, 4.0, 5.0, 6.0, -7.0},
 	},
 	{
